# sql/sqlbase: improve performance of fk checks