--- a/pkg/sql/bench_test.go
+++ b/pkg/sql/bench_test.go
@@ -272,12 +272,56 @@ func runBenchmarkInsert(b *testing.B, db *gosql.DB, count int) {
 
 }
 
+// runBenchmarkInsertFK benchmarks inserting count rows into a table with a
+// present foreign key into another table.
+func runBenchmarkInsertFK(b *testing.B, db *gosql.DB, count int) {
+	defer func() {
+		if _, err := db.Exec(`DROP TABLE IF EXISTS bench.insert`); err != nil {
+			b.Fatal(err)
+		}
+		if _, err := db.Exec(`DROP TABLE IF EXISTS bench.fk`); err != nil {
+			b.Fatal(err)
+		}
+	}()
+
+	if _, err := db.Exec(`CREATE TABLE bench.fk (k INT PRIMARY KEY)`); err != nil {
+		b.Fatal(err)
+	}
+	if _, err := db.Exec(`INSERT INTO bench.fk VALUES(1), (2), (3)`); err != nil {
+		b.Fatal(err)
+	}
+	if _, err := db.Exec(`CREATE TABLE bench.insert (k INT PRIMARY KEY, r INT, foreign key (r) references bench.fk(k))`); err != nil {
+		b.Fatal(err)
+	}
+
+	b.ResetTimer()
+	var buf bytes.Buffer
+	val := 0
+	for i := 0; i < b.N; i++ {
+		buf.Reset()
+		buf.WriteString(`INSERT INTO bench.insert VALUES `)
+		for j := 0; j < count; j++ {
+			if j > 0 {
+				buf.WriteString(", ")
+			}
+			fmt.Fprintf(&buf, "(%d, 1)", val)
+			val++
+		}
+		if _, err := db.Exec(buf.String()); err != nil {
+			b.Fatal(err)
+		}
+	}
+	b.StopTimer()
+
+}
+
 func BenchmarkSQL(b *testing.B) {
 	forEachDB(b, func(b *testing.B, db *gosql.DB) {
 		for _, runFn := range []func(*testing.B, *gosql.DB, int){
 			runBenchmarkDelete,
 			runBenchmarkInsert,
 			runBenchmarkInsertDistinct,
+			runBenchmarkInsertFK,
 			runBenchmarkInterleavedSelect,
 			runBenchmarkTrackChoices,
 			runBenchmarkUpdate,