There's no need to read or decode the full found fk row, or even decode
the found fk key. It's sufficient to just report whether there was any
key found in the span the fk should be in.

```
name                                 old time/op  new time/op  delta
SQL/Cockroach/InsertFK/count=1-8      451µs ± 1%   448µs ± 3%    ~     (p=0.684 n=10+10)
SQL/Cockroach/InsertFK/count=10-8    1.42ms ± 1%  1.40ms ± 2%  -1.42%  (p=0.015 n=8+9)
SQL/Cockroach/InsertFK/count=100-8   10.6ms ± 3%  10.3ms ± 3%  -2.74%  (p=0.001 n=10+9)
SQL/Cockroach/InsertFK/count=1000-8   105ms ± 3%   102ms ± 2%  -2.38%  (p=0.001 n=10+10)
```

This isn't a huge gain, but the larger the rows on the foreign table are, and the more column families they have, the larger the gain will be.