@nvanbenschoten 2017-09-19
Nice find!

---

Reviewed 2 of 2 files at r1, 1 of 1 files at r2.
Review status: all files reviewed at latest revision, 3 unresolved discussions, some commit checks pending.

---

*[pkg/sql/bench_test.go, line 275 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/18591#-KuPzn5yfCLWKrZbLoIu:-KuPzn5yfCLWKrZbLoIv:buq61g0) ([raw file](https://github.com/cockroachdb/cockroach/blob/0a03e59134d6279567b211e644c1f077c4785662/pkg/sql/bench_test.go#L275)):*
> ```Go
> }
> 
> // runBenchmarkInsert benchmarks inserting count rows into a table.
> ```

`s/runBenchmarkInsert/runBenchmarkInsertFK`. Also, this comment should be different from the one above on `runBenchmarkInsert`.

---

*[pkg/sql/sqlbase/fk.go, line 329 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/18591#-KuQ-kiyE-TjenjMw0NO:-KuQ-kiyE-TjenjMw0NP:b-rqtptb) ([raw file](https://github.com/cockroachdb/cockroach/blob/2f563ebdd91f89172e77f4dbd8d9f0d55c739e62/pkg/sql/sqlbase/fk.go#L329)):*
> ```Go
> 	b.searchIdx = searchIdx
> 	ids := ColIDtoRowIndexFromCols(b.searchTable.Columns)
> 	needed := make([]bool, len(ids))
> ```

Do we even need to allocate this? Can we just pass in `nil`?

---

*[pkg/sql/sqlbase/fk.go, line 355 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/18591#-KuQ-YElHGs4a58gz5EL:-KuQ-YElHGs4a58gz5EM:b-hno7h0) ([raw file](https://github.com/cockroachdb/cockroach/blob/2f563ebdd91f89172e77f4dbd8d9f0d55c739e62/pkg/sql/sqlbase/fk.go#L355)):*
> ```Go
> 
> // TODO(dt): Batch checks of many rows.
> func (f baseFKHelper) check(ctx context.Context, values parser.Datums, traceKV bool) (bool, error) {
> ```

Add a comment to this method.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/18591#-:-KuQ-t8FyoGXBHeg1WpT:b-a5n7j4)*
<!-- Sent from Reviewable.io -->