@jordanlewis 2017-09-19
Review status: all files reviewed at latest revision, 3 unresolved discussions, some commit checks pending.

---

*[pkg/sql/bench_test.go, line 275 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/18591#-KuPzn5yfCLWKrZbLoIu:-KuQ0cuflu9P7gEpoUvC:bln6nrr) ([raw file](https://github.com/cockroachdb/cockroach/blob/0a03e59134d6279567b211e644c1f077c4785662/pkg/sql/bench_test.go#L275)):*
<details><summary><i>Previously, nvanbenschoten (Nathan VanBenschoten) wrote…</i></summary><blockquote>

`s/runBenchmarkInsert/runBenchmarkInsertFK`. Also, this comment should be different from the one above on `runBenchmarkInsert`.
</blockquote></details>

D'oh, done.

---

*[pkg/sql/sqlbase/fk.go, line 329 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/18591#-KuQ-kiyE-TjenjMw0NO:-KuQ1WZItsuk9vMo1FAR:bfbq99i) ([raw file](https://github.com/cockroachdb/cockroach/blob/2f563ebdd91f89172e77f4dbd8d9f0d55c739e62/pkg/sql/sqlbase/fk.go#L329)):*
<details><summary><i>Previously, nvanbenschoten (Nathan VanBenschoten) wrote…</i></summary><blockquote>

Do we even need to allocate this? Can we just pass in `nil`?
</blockquote></details>

Good point, done.

---

*[pkg/sql/sqlbase/fk.go, line 355 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/18591#-KuQ-YElHGs4a58gz5EL:-KuQ0eHSvU78NckC54jg:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/2f563ebdd91f89172e77f4dbd8d9f0d55c739e62/pkg/sql/sqlbase/fk.go#L355)):*
<details><summary><i>Previously, nvanbenschoten (Nathan VanBenschoten) wrote…</i></summary><blockquote>

Add a comment to this method.
</blockquote></details>

Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/18591)*
<!-- Sent from Reviewable.io -->